Pull icons into its own component
